requirements.txt
qianfan
appbuilder-sdk
pyyaml